Split impl and trait in semantic.
