+++ crates/semantic/src/items/mod.rs
@@ -5,6 +5,7 @@ pub mod extern_type;
 pub mod free_function;
 pub mod functions;
 pub mod generics;
+pub mod imp;
 pub mod modifiers;
 pub mod strct;
 pub mod trt;
